requirements.txt
python-dotenv==1.0.1
simple_term_menu==1.6.4
